requirements.txt
pygame-ce==2.5.7
dearpygui==2.2
numpy==2.4.2